removed bulma; App & TeamCard in folders; sass re-org
