+++ package.json
@@ -6,7 +6,6 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "bulma": "^0.8.2",
     "csv-parser": "^2.3.2",
     "node-sass": "^4.14.0",
     "react": "^16.13.1",
