+++ src/App/App.scss
@@ -1,0 +1,7 @@
+.App {
+  display: flex;
+
+  & > div {
+    width: 45%;
+  }
+}
