+++ src/TeamCard/index.js
@@ -1,0 +1,2 @@
+import TeamCard from './TeamCard';
+export default TeamCard;
